refactor: コンポーネントと型の配置整理
